Add WIP version of d3 visualizer.js
Visualizes commit diff stats.
For example the file structure changes as you walk commit by commit.

This is a work in progress, it still needs the HTML and css to be checked in. Mostly here to document my work. Code is influenced heavily by previous yammer hackathon and nilclass.com work. The visualization technically works but you'll need the CSS and #world DOM node in the page first.